+++ config/dev.exs
@@ -2,6 +2,7 @@ import Config
 
 config :openpay,
   client_secret: "sk_da05dc98605d4e1d880582b4ee621e84",
+  client_public: "pk_8ffec93a697248e881cd4f67d027f81a",
   api_env: :sandbox,
   merchant_id: "mjtkrswiqtxftemz4tgl",
   webhook_auth: [